Update staff_ghosts.h (#176)

+++ src/staff_ghosts.h
@@ -40,7 +40,7 @@ extern s32 D_80162DF4;
 extern s32 D_80162DF8;
 
 extern u8* sReplayGhostBuffer;
-extern s16 sReplayGhostBufferSize;
+extern size_t sReplayGhostBufferSize;
 extern u32* sReplayGhostDecompressed;
 
 #endif /* STAFF_GHOSTS_H */
